renamed some report functions
