+++ oya-web/src/report/ReportPage.tsx
@@ -3,33 +3,20 @@ import React from 'react';
 import { Route, Routes, useLocation } from 'react-router';
 import { Link } from 'react-router-dom';
 import { Activity } from '../apiService';
-import BarReportTotal from './BarReportTotal';
-import PieReportTotal from './PieReportTotal';
+import BarTotalRechart from './BarTotalRechart';
+import PieTotalRechart from './PieTotalRechart';
 import ReportProvider, { useReportContext } from './ReportProvider';
-import { createActivityTotalReport, createDailyDataMap } from './reportUtils';
-import TreemapReportReactVis from './TreemapReportReactVis';
-import TreemapReportRechart from './TreemapReportRechart';
-import useActivities from '../useActivities';
-import useIntervals from '../useIntervals';
+import TreemapReactVis from './TreemapReactVis';
+import TreemapRechart from './TreemapRechart';
 import { getDeltaStringOfRange as ts } from '../utils';
 import TreemapNivo from './TreemapNivo';
 
-export function useReport() {
-  const { intervals } = useIntervals();
-  const { activityMappings, activities } = useActivities();
-  const ddm = createDailyDataMap(intervals, activityMappings);
-  const dda = Object.values(ddm);
-  const atrm = createActivityTotalReport(dda);
-  const atra = Object.values(atrm).sort((a, b) => Number(b?.time) - Number(a?.time));
-  return { intervals, activityMappings, activities, ddm, dda, atrm, atra };
-}
-
 const reportRoutes = [
   { link: 'cards', label: 'cards', element: <GripOverviewReport /> },
-  { link: 'bar', label: 'bars', element: <BarReportTotal /> },
-  { link: 'pie', label: 'pie', element: <PieReportTotal /> },
-  { link: 'vis-tree', label: 'vis tree', element: <TreemapReportReactVis /> },
-  { link: 're-tree', label: 'recharts tree', element: <TreemapReportRechart /> },
+  { link: 'bar', label: 'bars', element: <BarTotalRechart /> },
+  { link: 'pie', label: 'pie', element: <PieTotalRechart /> },
+  { link: 'vis-tree', label: 'vis tree', element: <TreemapReactVis /> },
+  { link: 're-tree', label: 'recharts tree', element: <TreemapRechart /> },
   { link: 'nivo-tree', label: 'nivo tree', element: <TreemapNivo /> },
 ];
